clean: move barely used utils function to the only file where it is used
We might rehabilitate it in the future, but it would be better if
it also worked with arrays.
